+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.12.0+5
+
+* **✨ New:** Added `CrashReportType` to control the leading information that gets sent to your crash reports. Configure it through the `Loglytics.setUp` method. Defaults to `location` (like before).
+* Moved icon and tag to beginning of log.
+
 ## 0.12.0+4
 
 * **✨ New:** Added `unverified` option to the `AnalyticsService`.